refactor: move the pull-request and review-style modules out of dashboard
Eight modules under agent/dashboard/ are named for the UI that edits them
rather than for what they are. Four talk to the GitHub pull-request API and
import agent.github.http directly; four are the review-style and eval domain
and import agent.review.

  dashboard/pull_request_context.py -> github/pull_request_context.py
  dashboard/pull_request_status.py  -> github/pull_request_status.py
  dashboard/pull_request_checks.py  -> github/pull_request_checks.py
  dashboard/pr_diff.py              -> github/pull_request_diff.py
  dashboard/review_styles.py        -> review/styles.py
  dashboard/review_style_jobs.py    -> review/style_jobs.py
  dashboard/analyzer_cron.py        -> review/analyzer_cron.py
  dashboard/eval_jobs.py            -> review/eval_jobs.py

pr_diff is renamed for consistency with the three modules it now sits beside.

agent/github/ now holds both checks.py and pull_request_checks.py. They are
different: checks.py reads one ref's check runs and statuses for CI triage;
pull_request_checks.py batches check state for many pull requests in a single
GraphQL round trip for the thread list.
