+++ agent/dashboard/environments.py
@@ -27,7 +27,7 @@ from typing import Any, Literal, TypedDict
 
 from pydantic import BaseModel, ConfigDict, Field, JsonValue, field_validator
 
-from agent.dashboard.review_styles import normalize_repo_full_name
+from agent.review.styles import normalize_repo_full_name
 from agent.store import TypedStore, now_iso
 
 logger = logging.getLogger(__name__)
