+++ agent/dashboard/review_api.py
@@ -17,9 +17,9 @@ from urllib.parse import urljoin, urlparse
 import httpx
 from fastapi import HTTPException, Response
 
-from agent.dashboard.pr_diff import build_pr_diff_files
 from agent.github.app import get_github_app_installation_token
 from agent.github.checks import github_headers
+from agent.github.pull_request_diff import build_pr_diff_files
 from agent.github.webhook import trigger_pr_review_from_ref
 from agent.review.findings import (
     REVIEWER_THREAD_KIND,
