+++ agent/review/analyzer_cron.py
@@ -9,12 +9,12 @@ token resolved inside ``get_analyzer`` (the cron carries no fresh user token).
 import hashlib
 import logging
 
-from agent.dashboard.review_style_jobs import (
+from agent.review.style_jobs import (
     _client,
     build_continual_run_configurable,
     build_continual_run_input,
 )
-from agent.dashboard.review_styles import REVIEW_STYLES
+from agent.review.styles import REVIEW_STYLES
 
 logger = logging.getLogger(__name__)
