+++ agent/review/style_jobs.py
@@ -7,16 +7,16 @@ from typing import Any
 
 from langgraph_sdk import get_client
 
-from agent.dashboard.review_styles import (
-    REVIEW_STYLES,
-    ReviewStyle,
-    reconcile_running_status,
-)
 from agent.dispatch import create_durable_run
 from agent.input_messages import RunInput, build_run_input
 from agent.review.style_collector import (
     collect_review_samples,
     format_samples_for_analyzer,
+)
+from agent.review.styles import (
+    REVIEW_STYLES,
+    ReviewStyle,
+    reconcile_running_status,
 )
 from agent.thread_ids import review_style_thread_id
 from agent.utils.analyzer_skills import build_skill_files
