+++ tests/dashboard/test_pull_request_checks.py
@@ -1,5 +1,5 @@
-import agent.dashboard.pull_request_checks as checks_module
-from agent.dashboard.pull_request_checks import get_pull_request_check_states
+import agent.github.pull_request_checks as checks_module
+from agent.github.pull_request_checks import get_pull_request_check_states
 
 
 def _rollup(state: str | None, pr_state: str = "OPEN", is_draft: bool = False) -> dict[str, object]:
